Maintain Order from seller and buyer updated

+++ angular-services/controllers/order-controller.js
@@ -33,12 +33,19 @@ exports.list = (req, res, next) => {
         });
 }
 
-exports.deleteById = (req, res, next) => {
-    Order.findByIdAndDelete(req.params.orderId)
-        .then(result => {
-            res.status(200).send(result);
-        })
-        .catch(err => {
-            res.status(500).send({ errMsg: err });
-        });
+exports.cancelById = async (req, res, next) => {
+    const orderStatus = await Order.find({$and:[{"_id":mongoose.Types.ObjectId(req.params.orderId)},{"status":"Pending"}]});
+    if(orderStatus.length===0){
+        res.status(401).send({
+            errors: { "Cannot cancel this order": ["It has already been shipped."] },
+          }); 
+    } else {
+        Order.findByIdAndDelete(req.params.orderId)
+            .then(result => {
+                res.status(200).send({});
+            })
+            .catch(err => {
+                res.status(500).send({ errMsg: err });
+            });
+    }
 }
